Move root project route to its own component
